improve code maintainability
